+++ sdk/inc/azure/iot/az_iot_provisioning_client.h
@@ -78,7 +78,9 @@ AZ_NODISCARD az_iot_provisioning_client_options az_iot_provisioning_client_optio
  * part of the certificate subject). Must conform to the limitations listed in the link below:
  * https://docs.microsoft.com/azure/iot-dps/concepts-service#registration-id
  * @param[in] options __[nullable]__ A reference to an #az_iot_provisioning_client_options
- * structure. Can be `NULL` for default options.
+ * structure. If `NULL` is passed, the provisioning client will use the default options. If using
+ * custom options, please initialize first by calling az_iot_provisioning_client_options_default()
+ * and then populating relevant options with your own values.
  * @pre \p client must not be `NULL`.
  * @pre \p global_device_hostname must be a valid span of size greater than 0.
  * @pre \p id_scope must be a valid span of size greater than 0.
@@ -271,6 +273,12 @@ typedef struct
    * Submit this timestamp when asking for Azure IoT service-desk help.
    */
   az_span error_timestamp;
+
+  /**
+   * An optional custom payload received from the service.
+   */
+  az_span payload;
+
 } az_iot_provisioning_client_registration_state;
 
 /**
@@ -440,7 +448,8 @@ AZ_NODISCARD az_result az_iot_provisioning_client_query_status_get_publish_topic
 
 /**
  * @brief Azure IoT Provisioning Client options for
- * az_iot_provisioning_client_get_request_payload(). Not currently used.  Reserved for future use.
+ * az_iot_provisioning_client_get_request_payload() and
+ * az_iot_provisioning_client_register_get_request_payload().
  *
  */
 typedef struct
@@ -453,11 +462,61 @@ typedef struct
 } az_iot_provisioning_client_payload_options;
 
 /**
+ * @brief Gets the default az_iot_provisioning_client_payload_options.
+ * @details Call this to obtain an initialized #az_iot_provisioning_client_payload_options structure
+ * that can be afterwards modified and passed to
+ * #az_iot_provisioning_client_register_get_request_payload.
+ *
+ * @return #az_iot_provisioning_client_payload_options.
+ */
+AZ_NODISCARD az_iot_provisioning_client_payload_options
+az_iot_provisioning_client_payload_options_default();
+
+/**
+ * @brief Builds the optional payload for a provisioning request.
+ * @remark Use this API to build an MQTT payload during device registration.
+ *         This call is optional for most scenarios. Some service
+ *         applications may require `custom_payload_property` specified during
+ *         registration to make additional decisions during provisioning time.
+ *         For example, if you need to register an IoT Plug and Play device you must
+ *         specify its model_id with this API via the `custom_payload_property`
+ *         `{"modelId":"your_model_id"}`.
+ *
+ * @param[in] client The #az_iot_provisioning_client to use for this call.
+ * @param[in] custom_payload_property __[nullable]__ Custom JSON to be added to this payload.
+ * Can be `NULL`.
+ * @param[in] options A reference to an #az_iot_provisioning_client_payload_options
+ * structure. Must be initialized first by calling
+ * az_iot_provisioning_client_payload_options_default() and then populating relevant options with
+ * your own values.
+ * @param[out] mqtt_payload A buffer with sufficient capacity to hold the MQTT payload.
+ * @param[in] mqtt_payload_size The size, in bytes of \p mqtt_payload.
+ * @param[out] out_mqtt_payload_length Contains the length, in bytes, written to \p mqtt_payload on
+ * success.
+ * @pre \p client must not be `NULL`.
+ * @pre \p mqtt_payload must not be `NULL`.
+ * @pre \p mqtt_payload_size must be greater than 0.
+ * @pre \p out_mqtt_payload_length must not be `NULL`.
+ * @return An #az_result value indicating the result of the operation.
+ * @retval #AZ_OK The payload was created successfully.
+ * @retval #AZ_ERROR_NOT_ENOUGH_SPACE The buffer is too small.
+ */
+AZ_NODISCARD az_result az_iot_provisioning_client_register_get_request_payload(
+    az_iot_provisioning_client const* client,
+    az_span custom_payload_property,
+    az_iot_provisioning_client_payload_options const* options,
+    uint8_t* mqtt_payload,
+    size_t mqtt_payload_size,
+    size_t* out_mqtt_payload_length);
+
+/**
+ * @deprecated since 1.5.0-beta.1.
+ * @see az_iot_provisioning_client_register_get_request_payload
  * @brief Builds the optional payload for a provisioning request.
  * @remark Use this API to build an MQTT payload during registration.
  *         This call is optional for most scenarios. Some service
  *         applications may require `custom_payload_property` specified during
- *         registration to take additional decisions during provisioning time.
+ *         registration to make additional decisions during provisioning time.
  *         For example, if you need to register an IoT Plug and Play device you must
  *         specify its model_id with this API via the `custom_payload_property`
  *         `{"modelId":"your_model_id"}`.
@@ -480,7 +539,7 @@ typedef struct
  * @retval #AZ_OK The payload was created successfully.
  * @retval #AZ_ERROR_NOT_ENOUGH_SPACE The buffer is too small.
  */
-AZ_NODISCARD az_result az_iot_provisioning_client_get_request_payload(
+AZ_DEPRECATED AZ_NODISCARD az_result az_iot_provisioning_client_get_request_payload(
     az_iot_provisioning_client const* client,
     az_span custom_payload_property,
     az_iot_provisioning_client_payload_options const* options,